feature: start AuthService

+++ Blog.API/ViewModels/AuthData.cs
@@ -3,7 +3,7 @@ namespace Blog.API.ViewModels
     public class AuthData
     {
         public string Token { get; set; }
-        public int TokenExpirationTime { get; set; }
+        public long TokenExpirationTime { get; set; }
         public int Id { get; set; }
     }
 }
